Fixed import issue by combining common code from the 2 main files into a util file
